Require webworker for lean-client-js-browser
(This cuts the browser bundle size in half!)
