+++ lean-client-js-browser/lib_test.html
@@ -19,10 +19,7 @@
         libraryZip: prefix+'/library.zip',
     };
 
-    const transport =
-        window.Worker ?
-            new lean.WebWorkerTransport(opts) :
-            new lean.BrowserInProcessTransport(opts);
+    const transport = new lean.WebWorkerTransport(opts);
     const server = new lean.Server(transport);
     server.error.on((err) => console.log('error:', err));
     server.allMessages.on((allMessages) => console.log('messages:', allMessages.msgs));
