+++ lean-client-js-browser/src/inprocesstypes.ts
@@ -1,0 +1,12 @@
+export interface LeanJsUrls {
+    libraryZip?: string;
+
+    javascript?: string;
+
+    webassemblyWasm?: string;
+    webassemblyJs?: string;
+}
+
+export interface LeanJsOpts extends LeanJsUrls {
+    memoryMB?: number;
+}
